use JWT during apiClient instantiation
