+++ src/stores/useMessages.ts
@@ -2,7 +2,7 @@ import type { MessagesState } from '@stores/types'
 import type { BotMessage, UserMessage } from '@models/Message'
 import { uniqueId } from 'lodash'
 import { useNotifications } from '@stores/useNotifications'
-import { apiClient, updateCredential } from '@/api'
+import { apiClient } from '@/api'
 import MemoryService from '@services/MemoryService'
 import { useSettings } from './useSettings'
 
@@ -51,7 +51,6 @@ export const useMessages = defineStore('messages', () => {
 	})
 
 	const { showNotification } = useNotifications()
-	const { cookie } = storeToRefs(useSettings())
 
 	watchEffect(() => {
 		/**
@@ -113,8 +112,6 @@ export const useMessages = defineStore('messages', () => {
 			.onDisconnected(() => {
 				currentState.ready = false
 			})
-
-		updateCredential(cookie.value)
 	})
 
 	tryOnUnmounted(() => {
